use original training

+++ src/fl-client-ti.py
@@ -35,8 +35,8 @@ warnings.filterwarnings("ignore", category=UserWarning)
 def load_data(cid:int=0, n_splits:int=1, timesteps:int=10):
 
     data_directory = "canarybit-heflp/data/ti_training" 
-    print("CLIENT CID", str(cid/2), cid%2)
-    input_data_folder_path=os.path.join(os.path.dirname(os.getcwd()), data_directory, str(cid/2))
+    print("CLIENT CID", str(int(cid/2)), cid%2)
+    input_data_folder_path=os.path.join(os.path.dirname(os.getcwd()), data_directory, str(int(cid/2)))
 
     X_train = reading_files(os.path.join(input_data_folder_path,"X_train_formatted.csv"))
     X_test = reading_files(os.path.join(input_data_folder_path,"X_test_formatted.csv"))
